refactor: refine app structure
