+++ src/index.ts
@@ -1,8 +1,9 @@
 import { Hono } from "hono";
 import { cors } from "hono/cors";
 import { logger } from "hono/logger";
-import { rootRoute } from "./roots";
-import { provinceRoute } from "./provinces";
+
+import { rootRoute } from "./routes/roots";
+import { provinceRoute } from "./routes/provinces";
 
 const app = new Hono();
